+++ oo/src/language/analyzing/types.rs
@@ -13,6 +13,7 @@ pub enum VariableType {
     CHAR,
     STRING,
     NULL,
+    RANGE,
 }
 
 impl VariableType {